+++ src/main/java/com/project/bookstore/service/BookService.java
@@ -7,11 +7,11 @@ import com.project.bookstore.entities.Book;
 public interface BookService {
 	public List<Book> showAllBook();
 
-	public void deleteBookById(int id);
+	public void deleteBookById(long id);
 
 	public void addOrUpdateBook(Book book);
 
-	public Book findBookById(int id);
+	public Book findBookById(long id);
 	
 
 }
